First version of the AWS Config to Slack Bridge
This adds the first version of the AWS Config to Slack bridge. It is a lambda that runs daily and pulls identified resources from an AWS Config Aggregator and reports on them towards slack.

To avoid showing just anonymous account IDs it integrates with AWS Organizations to resolve them to account names.